+++ engine/src/flutter/impeller/display_list/aiks_dl_unittests.cc
@@ -19,9 +19,15 @@
 #include "flutter/display_list/dl_color.h"
 #include "flutter/display_list/dl_paint.h"
 #include "flutter/testing/testing.h"
+#include "fml/synchronization/count_down_latch.h"
 #include "imgui.h"
+#include "impeller/core/device_buffer.h"
+#include "impeller/core/device_buffer_descriptor.h"
+#include "impeller/core/formats.h"
+#include "impeller/core/texture_descriptor.h"
 #include "impeller/display_list/dl_dispatcher.h"
 #include "impeller/display_list/dl_image_impeller.h"
+#include "impeller/geometry/path_builder.h"
 #include "impeller/geometry/scalar.h"
 #include "include/core/SkCanvas.h"
 #include "include/core/SkMatrix.h"
@@ -1040,6 +1046,80 @@ TEST_P(AiksTest, DepthValuesForPolygonMode) {
   builder.Restore();
 
   ASSERT_TRUE(OpenPlaygroundHere(builder.Build()));
+}
+
+// Verifies that an image rasterized and readback is in the correct orientation
+// by re-uploading it.
+TEST_P(AiksTest, ToImageFromImage) {
+  DisplayListBuilder builder;
+  Path ip_path = PathBuilder{}
+                     .AddArc(DlRect::MakeLTRB(0, 0, 100, 100), Radians(0),
+                             Radians(3.14 / 2))
+                     .TakePath();
+  DlPath path = DlPath(ip_path);
+
+  builder.DrawPath(path, DlPaint().setColor(DlColor::kRed()));
+
+  AiksContext renderer(GetContext(), nullptr);
+  auto texture =
+      DisplayListToTexture(builder.Build(), ISize(100, 100), renderer);
+
+  // First, Readback the texture data into a host buffer.
+  impeller::DeviceBufferDescriptor desc;
+  desc.size = texture->GetTextureDescriptor().GetByteSizeOfBaseMipLevel();
+  desc.readback = true;
+  desc.storage_mode = StorageMode::kHostVisible;
+
+  auto device_buffer = GetContext()->GetResourceAllocator()->CreateBuffer(desc);
+  {
+    auto cmd_buffer = GetContext()->CreateCommandBuffer();
+    auto blit_pass = cmd_buffer->CreateBlitPass();
+
+    blit_pass->AddCopy(texture, device_buffer);
+    blit_pass->EncodeCommands();
+
+    auto latch = std::make_shared<fml::CountDownLatch>(1u);
+    GetContext()->GetCommandQueue()->Submit(
+        {cmd_buffer},
+        [latch](CommandBuffer::Status status) { latch->CountDown(); });
+    latch->Wait();
+  }
+
+  impeller::TextureDescriptor tex_desc = texture->GetTextureDescriptor();
+  auto reupload_texture =
+      GetContext()->GetResourceAllocator()->CreateTexture(tex_desc);
+
+  // Next, Re-upload the data into a new texture.
+  {
+    auto cmd_buffer = GetContext()->CreateCommandBuffer();
+    auto blit_pass = cmd_buffer->CreateBlitPass();
+    blit_pass->AddCopy(DeviceBuffer::AsBufferView(device_buffer),
+                       reupload_texture);
+    blit_pass->ConvertTextureToShaderRead(texture);
+    blit_pass->EncodeCommands();
+
+    auto latch = std::make_shared<fml::CountDownLatch>(1u);
+    GetContext()->GetCommandQueue()->Submit(
+        {cmd_buffer},
+        [latch](CommandBuffer::Status status) { latch->CountDown(); });
+    latch->Wait();
+  }
+
+  // Draw the results side by side. These should look the same.
+  DisplayListBuilder canvas;
+  DlPaint paint = DlPaint();
+  canvas.DrawRect(
+      DlRect::MakeLTRB(0, 0, 100, 100),
+      DlPaint().setColor(DlColor::kBlue()).setDrawStyle(DlDrawStyle::kStroke));
+  canvas.DrawImage(DlImageImpeller::Make(texture), DlPoint(0, 0),
+                   DlImageSampling::kNearestNeighbor, &paint);
+
+  canvas.DrawRect(
+      DlRect::MakeLTRB(0, 100, 100, 200),
+      DlPaint().setColor(DlColor::kRed()).setDrawStyle(DlDrawStyle::kStroke));
+  canvas.DrawImage(DlImageImpeller::Make(reupload_texture), DlPoint(0, 100),
+                   DlImageSampling::kNearestNeighbor, &paint);
+  OpenPlaygroundHere(canvas.Build());
 }
 
 }  // namespace testing